+++ Core/SoarKernel/src/prefmem.cpp
@@ -86,7 +86,7 @@ preference *make_preference (agent* thisAgent, byte type, Symbol *id, Symbol *at
   p->wma_o_set = NIL;
 
 #ifdef DEBUG_PREFS
-  print (thisAgent, "\nAllocating preference at 0x%8x: ", (unsigned long)p);
+  print (thisAgent, "\nAllocating preference at 0x%8x: ", static_cast<uintptr_t>(p));
   print_preference (thisAgent, p);
 #endif
 
@@ -103,7 +103,7 @@ preference *make_preference (agent* thisAgent, byte type, Symbol *id, Symbol *at
 void deallocate_preference (agent* thisAgent, preference *pref) {
 
 #ifdef DEBUG_PREFS  
-  print (thisAgent, "\nDeallocating preference at 0x%8x: ",(unsigned long)pref);
+  print (thisAgent, "\nDeallocating preference at 0x%8x: ",static_cast<uintptr_t>(pref));
   print_preference (thisAgent, pref);
   if (pref->reference_count != 0) {   /* --- sanity check --- */
     char msg[BUFFER_MSG_SIZE];
@@ -214,7 +214,7 @@ void add_preference_to_tm (agent* thisAgent, preference *pref)
    preference *p2;
 
 #ifdef DEBUG_PREFS
-   print (thisAgent, "\nAdd preference at 0x%8x:  ",(unsigned long)pref);
+   print (thisAgent, "\nAdd preference at 0x%8x:  ",static_cast<uintptr_t>(pref));
    print_preference (thisAgent, pref);
 #endif
    
@@ -293,7 +293,7 @@ void remove_preference_from_tm (agent* thisAgent, preference *pref) {
   s = pref->slot;
 
 #ifdef DEBUG_PREFS
-  print (thisAgent, "\nRemove preference at 0x%8x:  ",(unsigned long)pref);
+  print (thisAgent, "\nRemove preference at 0x%8x:  ",static_cast<uintptr_t>(pref));
   print_preference (thisAgent, pref);
 #endif
 
@@ -345,7 +345,7 @@ void process_o_rejects_and_deallocate_them (agent* thisAgent, preference *o_reje
                                    a clone of some other pref we're about to
                                    remove */
 #ifdef DEBUG_PREFS
-  print (thisAgent, "\nO-reject posted at 0x%8x:  ",(unsigned long)pref);
+  print (thisAgent, "\nO-reject posted at 0x%8x:  ",static_cast<uintptr_t>(pref));
   print_preference (thisAgent, pref);
 #endif
   }
